// src/types/country.ts
export interface Country {
    name: Name
    tld?: string[]
    cca2: string
    ccn3?: string
    cca3: string
    cioc?: string
    independent?: boolean
    status: string
    unMember: boolean
    currencies?: Currencies
    idd: Idd
    capital?: string[]
    altSpellings: string[]
    region: string
    subregion?: string
    languages?: Languages
    translations: Translations
    latlng: number[]
    landlocked: boolean
    area: number
    demonyms?: Demonyms
    flag: string
    maps: Maps
    population: number
    fifa?: string
    car: Car
    timezones: string[]
    continents: string[]
    flags: Flags
    coatOfArms: CoatOfArms
    startOfWeek: string
    capitalInfo: CapitalInfo
    postalCode?: PostalCode
    borders?: string[]
    gini?: Gini
}

export interface Name {
    common: string
    official: string
    nativeName?: NativeName
}

export interface NativeName {
    eng?: Eng
    spa?: Spa
    hrv?: Hrv
    cha?: Cha
    ita?: Ita
    lat?: Lat
    ara?: Ara
    fra?: Fra
    pih?: Pih
    sqi?: Sqi
    srp?: Srp
    rus?: Rus
    tgk?: Tgk
    niu?: Niu
    fas?: Fas
    nrf?: Nrf
    khm?: Khm
    nau?: Nau
    bul?: Bul
    pau?: Pau
    por?: Por
    pov?: Pov
    fil?: Fil
    dan?: Dan
    fao?: Fao
    ell?: Ell
    swe?: Swe
    nld?: Nld
    arc?: Arc
    ckb?: Ckb
    swa?: Swa
    mya?: Mya
    som?: Som
    kor?: Kor
    smo?: Smo
    tkl?: Tkl
    mlt?: Mlt
    kin?: Kin
    tur?: Tur
    cal?: Cal
    bjz?: Bjz
    zho?: Zho
    msa?: Msa
    tam?: Tam
    mri?: Mri
    nzs?: Nzs
    kal?: Kal
    ber?: Ber
    hin?: Hin
    tha?: Tha
    nno?: Nno
    nob?: Nob
    smi?: Smi
    slk?: Slk
    kat?: Kat
    deu?: Deu
    ltz?: Ltz
    bar?: Bar
    sin?: Sin
    lao?: Lao
    mon?: Mon
    kon?: Kon
    lin?: Lin
    lua?: Lua
    mey?: Mey
    pol?: Pol
    jam?: Jam
    hmo?: Hmo
    tpi?: Tpi
    grn?: Grn
    pap?: Pap
    dzo?: Dzo
    tvl?: Tvl
    uzb?: Uzb
    sot?: Sot
    run?: Run
    vie?: Vie
    aym?: Aym
    que?: Que
    hye?: Hye
    kaz?: Kaz
    nor?: Nor
    isl?: Isl
    mfe?: Mfe
    gle?: Gle
    tsn?: Tsn
    hat?: Hat
    ukr?: Ukr
    hun?: Hun
    amh?: Amh
    tuk?: Tuk
    lit?: Lit
    ssw?: Ssw
    bos?: Bos
    cnr?: Cnr
    bel?: Bel
    nya?: Nya
    prs?: Prs
    pus?: Pus
    jpn?: Jpn
    glv?: Glv
    ind?: Ind
    tir?: Tir
    heb?: Heb
    ben?: Ben
    ron?: Ron
    nfr?: Nfr
    afr?: Afr
    her?: Her
    hgm?: Hgm
    kwn?: Kwn
    loz?: Loz
    ndo?: Ndo
    bis?: Bis
    cat?: Cat
    aze?: Aze
    div?: Div
    tet?: Tet
    mkd?: Mkd
    est?: Est
    ces?: Ces
    zdj?: Zdj
    rar?: Rar
    nep?: Nep
    mah?: Mah
    ton?: Ton
    gil?: Gil
    fin?: Fin
    urd?: Urd
    slv?: Slv
    kir?: Kir
    gsw?: Gsw
    roh?: Roh
    crs?: Crs
    fij?: Fij
    hif?: Hif
    sag?: Sag
    mlg?: Mlg
    nbl?: Nbl
    nso?: Nso
    tso?: Tso
    ven?: Ven
    xho?: Xho
    zul?: Zul
    lav?: Lav
    bwg?: Bwg
    kck?: Kck
    khi?: Khi
    ndc?: Ndc
    nde?: Nde
    sna?: Sna
    toi?: Toi
    zib?: Zib
}
// ...
export interface Currencies {
    USD?: Usd
    AUD?: Aud
    HNL?: Hnl
    SBD?: Sbd
    EUR?: Eur
    LBP?: Lbp
    OMR?: Omr
    GBP?: Gbp
    SHP?: Shp
    TJS?: Tjs
    NZD?: Nzd
    IRR?: Irr
    XOF?: Xof
    EGP?: Egp
    ILS?: Ils
    JOD?: Jod
    JEP?: Jep
    KHR?: Khr
    BGN?: Bgn
    SAR?: Sar
    PHP?: Php
    DKK?: Dkk
    FOK?: Fok
    SEK?: Sek
    SSP?: Ssp
    GHS?: Ghs
    RSD?: Rsd
    ANG?: Ang
    SLL?: Sll
    LYD?: Lyd
    XCD?: Xcd
    IQD?: Iqd
    TZS?: Tzs
    MMK?: Mmk
    LRD?: Lrd
    SOS?: Sos
    ALL?: All
    KRW?: Krw
    MXN?: Mxn
    TTD?: Ttd
    RWF?: Rwf
    TND?: Tnd
    AOA?: Aoa
    XAF?: Xaf
    BHD?: Bhd
    VES?: Ves
    FKP?: Fkp
    BZD?: Bzd
    SYP?: Syp
    SGD?: Sgd
    BRL?: Brl
    TWD?: Twd
    QAR?: Qar
    NIO?: Nio
    MAD?: Mad
    INR?: Inr
    THB?: Thb
    NOK?: Nok
    GEL?: Gel
    DOP?: Dop
    LKR?: Lkr
    KES?: Kes
    BND?: Bnd
    DZD?: Dzd
    LAK?: Lak
    MNT?: Mnt
    UGX?: Ugx
    CDF?: Cdf
    MRU?: Mru
    RUB?: Rub
    PLN?: Pln
    JMD?: Jmd
    PGK?: Pgk
    PYG?: Pyg
    AWG?: Awg
    BTN?: Btn
    TVD?: Tvd
    CHF?: Chf
    GMD?: Gmd
    KYD?: Kyd
    UZS?: Uzs
    GTQ?: Gtq
    BMD?: Bmd
    LSL?: Lsl
    ZAR?: Zar
    COP?: Cop
    CAD?: Cad
    BIF?: Bif
    XPF?: Xpf
    VND?: Vnd
    SDG?: Sdg
    BOB?: Bob
    AMD?: Amd
    MOP?: Mop
    CNY?: Cny
    KZT?: Kzt
    CLP?: Clp
    ISK?: Isk
    GIP?: Gip
    MUR?: Mur
    GYD?: Gyd
    BWP?: Bwp
    HTG?: Htg
    UAH?: Uah
    NGN?: Ngn
    HUF?: Huf
    ETB?: Etb
    TMT?: Tmt
    SZL?: Szl
    BAM?: Bam
    BYN?: Byn
    TRY?: Try
    MWK?: Mwk
    AED?: Aed
    AFN?: Afn
    JPY?: Jpy
    IMP?: Imp
    IDR?: Idr
    ERN?: Ern
    MZN?: Mzn
    BDT?: Bdt
    RON?: Ron2
    GGP?: Ggp
    ZWL?: Zwl
    MGA?: Mga
    ARS?: Ars
    BBD?: Bbd
    SRD?: Srd
    NAD?: Nad
    GNF?: Gnf
    VUV?: Vuv
    WST?: Wst
    AZN?: Azn
    MVR?: Mvr
    PAB?: Pab
    MKD?: Mkd2
    CZK?: Czk
    BSD?: Bsd
    UYU?: Uyu
    KMF?: Kmf
    CKD?: Ckd
    CRC?: Crc
    STN?: Stn
    HKD?: Hkd
    MYR?: Myr
    NPR?: Npr
    CUC?: Cuc
    CUP?: Cup
    KPW?: Kpw
    MDL?: Mdl
    ZMW?: Zmw
    TOP?: Top
    PEN?: Pen
    CVE?: Cve
    KID?: Kid
    PKR?: Pkr
    DJF?: Djf
    KGS?: Kgs
    KWD?: Kwd
    SCR?: Scr
    FJD?: Fjd
    YER?: Yer
}

export interface Usd {
    name: string
    symbol: string
}
export interface Currency {
    name: string
    symbol: string
}
export interface Aud {
    name: string
    symbol: string
}

export interface Hnl {
    name: string
    symbol: string
}

export interface Sbd {
    name: string
    symbol: string
}

export interface Eur {
    name: string
    symbol: string
}

export interface Lbp {
    name: string
    symbol: string
}

export interface Omr {
    name: string
    symbol: string
}

export interface Gbp {
    name: string
    symbol: string
}

export interface Shp {
    name: string
    symbol: string
}

export interface Tjs {
    name: string
    symbol: string
}

export interface Nzd {
    name: string
    symbol: string
}

export interface Irr {
    name: string
    symbol: string
}

export interface Xof {
    name: string
    symbol: string
}

export interface Egp {
    name: string
    symbol: string
}

export interface Ils {
    name: string
    symbol: string
}

export interface Jod {
    name: string
    symbol: string
}

export interface Jep {
    name: string
    symbol: string
}

export interface Khr {
    name: string
    symbol: string
}

export interface Bgn {
    name: string
    symbol: string
}

export interface Sar {
    name: string
    symbol: string
}

export interface Php {
    name: string
    symbol: string
}

export interface Dkk {
    name: string
    symbol: string
}

export interface Fok {
    name: string
    symbol: string
}

export interface Sek {
    name: string
    symbol: string
}

export interface Ssp {
    name: string
    symbol: string
}

export interface Ghs {
    name: string
    symbol: string
}

export interface Rsd {
    name: string
    symbol: string
}

export interface Ang {
    name: string
    symbol: string
}

export interface Sll {
    name: string
    symbol: string
}

export interface Lyd {
    name: string
    symbol: string
}

export interface Xcd {
    name: string
    symbol: string
}

export interface Iqd {
    name: string
    symbol: string
}

export interface Tzs {
    name: string
    symbol: string
}

export interface Mmk {
    name: string
    symbol: string
}

export interface Lrd {
    name: string
    symbol: string
}

export interface Sos {
    name: string
    symbol: string
}

export interface All {
    name: string
    symbol: string
}

export interface Krw {
    name: string
    symbol: string
}

export interface Mxn {
    name: string
    symbol: string
}

export interface Ttd {
    name: string
    symbol: string
}

export interface Rwf {
    name: string
    symbol: string
}

export interface Tnd {
    name: string
    symbol: string
}

export interface Aoa {
    name: string
    symbol: string
}

export interface Xaf {
    name: string
    symbol: string
}

export interface Bhd {
    name: string
    symbol: string
}
// ...
export interface Idd {
    root?: string
    suffixes?: string[]
}

export interface Languages {
    eng?: string
    spa?: string
    hrv?: string
    cha?: string
    ita?: string
    lat?: string
    ara?: string
    fra?: string
    pih?: string
    sqi?: string
    srp?: string
    rus?: string
    tgk?: string
    niu?: string
    fas?: string
    nrf?: string
    khm?: string
    nau?: string
    bul?: string
    pau?: string
    por?: string
    pov?: string
    fil?: string
    dan?: string
    fao?: string
    ell?: string
    swe?: string
    nld?: string
    arc?: string
    ckb?: string
    swa?: string
    mya?: string
    som?: string
    kor?: string
    smo?: string
    tkl?: string
    mlt?: string
    kin?: string
    tur?: string
    cal?: string
    bjz?: string
    zho?: string
    msa?: string
    tam?: string
    mri?: string
    nzs?: string
    kal?: string
    ber?: string
    hin?: string
    tha?: string
    nno?: string
    nob?: string
    smi?: string
    slk?: string
    kat?: string
    deu?: string
    ltz?: string
    de?: string
    sin?: string
    lao?: string
    mon?: string
    kon?: string
    lin?: string
    lua?: string
    mey?: string
    pol?: string
    jam?: string
    hmo?: string
    tpi?: string
    grn?: string
    pap?: string
    dzo?: string
    tvl?: string
    uzb?: string
    sot?: string
    run?: string
    vie?: string
    aym?: string
    que?: string
    hye?: string
    kaz?: string
    nor?: string
    isl?: string
    mfe?: string
    gle?: string
    tsn?: string
    hat?: string
    ukr?: string
    hun?: string
    amh?: string
    tuk?: string
    lit?: string
    ssw?: string
    bos?: string
    cnr?: string
    bel?: string
    nya?: string
    prs?: string
    pus?: string
    jpn?: string
    glv?: string
    ind?: string
    tir?: string
    heb?: string
    ben?: string
    ron?: string
    nfr?: string
    afr?: string
    her?: string
    hgm?: string
    kwn?: string
    loz?: string
    ndo?: string
    bis?: string
    cat?: string
    aze?: string
    div?: string
    tet?: string
    mkd?: string
    est?: string
    ces?: string
    zdj?: string
    rar?: string
    nep?: string
    mah?: string
    ton?: string
    gil?: string
    fin?: string
    urd?: string
    slv?: string
    kir?: string
    gsw?: string
    roh?: string
    crs?: string
    fij?: string
    hif?: string
    sag?: string
    mlg?: string
    nbl?: string
    nso?: string
    tso?: string
    ven?: string
    xho?: string
    zul?: string
    lav?: string
    bwg?: string
    kck?: string
    khi?: string
    ndc?: string
    nde?: string
    sna?: string
    toi?: string
    zib?: string
}

export interface Translations {
    ara: Ara2
    bre: Bre
    ces: Ces2
    cym: Cym
    deu: Deu2
    est: Est2
    fin: Fin2
    fra: Fra2
    hrv?: Hrv2
    hun: Hun2
    ita: Ita2
    jpn?: Jpn2
    kor: Kor2
    nld: Nld2
    per?: Per
    pol: Pol2
    por: Por2
    rus: Rus2
    slk: Slk2
    spa: Spa2
    srp: Srp2
    swe: Swe2
    tur: Tur2
    urd: Urd2
    zho?: Zho2
}
// ...
export interface Demonyms {
    eng: Eng2
    fra?: Fra3
}

export interface Eng2 {
    f: string
    m: string
}

export interface Fra3 {
    f: string
    m: string
}

export interface Maps {
    googleMaps: string
    openStreetMaps: string
}

export interface Car {
    signs?: string[]
    side: string
}

export interface Flags {
    png: string
    svg: string
    alt?: string
}

export interface CoatOfArms {
    png?: string
    svg?: string
}

export interface CapitalInfo {
    latlng?: number[]
}

export interface PostalCode {
    format: string
    regex?: string
}

export interface Gini {
    "2019"?: number
    "2012"?: number
    "2018"?: number
    "2011"?: number
    "2017"?: number
    "2015"?: number
    "2009"?: number
    "2016"?: number
    "2010"?: number
    "2014"?: number
    "1992"?: number
    "2006"?: number
    "1999"?: number
    "2003"?: number
    "2013"?: number
    "2004"?: number
    "1998"?: number
    "2008"?: number
    "2005"?: number
}